warnings on associationView, matches and view all in player profile.
