+++ src/modules/athleteprofile/playerProfile.js
@@ -173,7 +173,7 @@ const matches = [
     journey: 5,
     date: `${7}/${10}/${2020}`,
     time: `${7}/${11}/${2020},${14}:${11}`,
-    estimateEndTime: `${16}:${50}`,
+    estimateEndTime: `${7}/${11}/${2020},${16}:${50}`,
     stadium: 'Estadio X',
     price: 'Q25.00',
     parking: 'Sí, gratis',
@@ -188,7 +188,7 @@ const matches = [
       },
     ],
     event: {
-      id: 1,
+      id: 9,
       title: 'Torneo 2021',
       category: 'Masculino Mayor A',
       winner: '',
@@ -280,7 +280,7 @@ const matches = [
     journey: 5,
     date: `${7}/${11}/${2020}`,
     time: `${7}/${11}/${2020},${14}:${11}`,
-    estimateEndTime: '16:00',
+    estimateEndTime: `${7}/${11}/${2020},${16}:00`,
     stadium: 'Estadio X',
     price: 'Q25.00',
     parking: 'Sí, gratis',
