+++ src/modules/matchs/matchbase/matchBase.js
@@ -4,6 +4,7 @@ import {
   StyleSheet, View, Text, Pressable,
 } from 'react-native'
 import TeamCircle from '../../shared/teamCircle/teamCircle'
+import { dateFormat, timeFormat } from '../../../utils/formats'
 
 const styles = StyleSheet.create({
   contentContainer: {
@@ -31,50 +32,32 @@ const styles = StyleSheet.create({
 
 const MatchBase = ({
   type, match, event, navigation,
-}) => {
-  const dateFormat = (date) => {
-    const dateObj = new Date(date)
-    const day = dateObj.getDate()
-    const month = dateObj.getMonth() + 1
-    const year = dateObj.getFullYear()
-    return `${day}/${month}/${year}`
-  }
-  const timeFormat = (date) => {
-    // ISO 8601 to local time
-    const dateObj = new Date(date)
-    const hours = dateObj.getUTCHours()
-    const minutes = dateObj.getUTCMinutes()
-    // create time string with leading zeros
-    const time = `${hours < 10 ? `0${hours}` : hours}:${minutes < 10 ? `0${minutes}` : minutes}`
-    return time
-  }
-  return (
-    <Pressable
-      style={({ pressed }) => [
-        (pressed) ? styles.pressed : {},
-        styles.contentContainer,
-      ]}
-      onPress={() => navigation.navigate('Match Description', {
-        event,
-        match,
-        type,
-      })}
-    >
-      <Text>{`${event.title} - ${event.category} - Jornada: ${match.journey}`}</Text>
-      <View style={styles.versusContainer}>
-        <TeamCircle team={match.local} />
-        <Text style={styles.result}>
-          {
-              (match.localScore === null) ? 'VS' : `${match.localScore} - ${match.visitScore}`
-            }
-        </Text>
-        <TeamCircle team={match.visit} side="right" />
-      </View>
-      <Text>{dateFormat(match.date)}</Text>
-      <Text>{timeFormat(match.time)}</Text>
-    </Pressable>
-  )
-}
+}) => (
+  <Pressable
+    style={({ pressed }) => [
+      (pressed) ? styles.pressed : {},
+      styles.contentContainer,
+    ]}
+    onPress={() => navigation.navigate('Match Description', {
+      event,
+      match,
+      type,
+    })}
+  >
+    <Text>{`${event.title} - ${event.category} - Jornada: ${match.journey}`}</Text>
+    <View style={styles.versusContainer}>
+      <TeamCircle team={match.local} />
+      <Text style={styles.result}>
+        {
+            (match.localScore === null) ? 'VS' : `${match.localScore} - ${match.visitScore}`
+          }
+      </Text>
+      <TeamCircle team={match.visit} side="right" />
+    </View>
+    <Text>{dateFormat(match.date)}</Text>
+    <Text>{timeFormat(match.time)}</Text>
+  </Pressable>
+)
 
 MatchBase.propTypes = {
   // eslint-disable-next-line react/forbid-prop-types
